requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile ./requirements.in
#
absl-py==2.1.0
    # via
    #   -r requirements.in
    #   tensorboard
    #   tensorflow
    #   tensorflow-probability
astunparse==1.6.3
    # via tensorflow
cachetools==5.3.3
    # via google-auth
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
cloudpickle==3.0.0
    # via tensorflow-probability
contextlib2==21.6.0
    # via -r requirements.in
decorator==5.1.1
    # via tensorflow-probability
dm-tree==0.1.8
    # via tensorflow-probability
flatbuffers==24.3.25
    # via tensorflow
gast==0.4.0
    # via
    #   tensorflow
    #   tensorflow-probability
google-auth==2.29.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==1.0.0
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.62.1
    # via
    #   tensorboard
    #   tensorflow
h5py==3.10.0
    # via tensorflow
idna==3.7
    # via requests
keras==2.13.1
    # via tensorflow
libclang==18.1.1
    # via tensorflow
lit==18.1.2
    # via -r requirements.in
markdown==3.6
    # via tensorboard
markupsafe==2.1.5
    # via werkzeug
networkx==3.2.1
    # via -r requirements.in
numpy==1.24.3
    # via
    #   -r requirements.in
    #   h5py
    #   opt-einsum
    #   scipy
    #   tensorboard
    #   tensorflow
    #   tensorflow-probability
oauthlib==3.2.2
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
packaging==24.0
    # via tensorflow
protobuf==4.25.3
    # via
    #   tensorboard
    #   tensorflow
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
requests==2.32.2
    # via
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
scipy==1.12.0
    # via -r requirements.in
semantic-version==2.10.0
    # via -r requirements.in
six==1.16.0
    # via
    #   astunparse
    #   google-pasta
    #   tensorflow
    #   tensorflow-probability
tensorboard==2.13.0
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.13.1 ; sys_platform == "linux"
    # via -r requirements.in
tensorflow-estimator==2.13.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.36.0
    # via tensorflow
tensorflow-probability==0.21.0
    # via -r requirements.in
termcolor==2.4.0
    # via tensorflow
typing-extensions==4.5.0
    # via
    #   tensorflow
    #   tensorflow-probability
urllib3==2.2.2
    # via requests
werkzeug==3.0.1
    # via tensorboard
wheel==0.43.0
    # via
    #   astunparse
    #   tensorboard
wrapt==1.16.0
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
